+++ ClientTest/Client.cs
@@ -42,8 +42,12 @@ namespace ClientTest
             _Client.Keepalive.EnableTcpKeepAlives = true; 
             _Client.Settings.MutuallyAuthenticate = false;
             _Client.Settings.AcceptInvalidCertificates = true;
+            _Client.Settings.ConnectTimeoutMs = 5000;
+            // _Client.Settings.IdleServerTimeoutMs = 10000;
             _Client.Logger = Logger;
-            _Client.Connect();
+
+            // _Client.Connect();
+            _Client.ConnectWithRetries(5000);
 
             while (_RunForever)
             {