+++ SimpleTcp/SimpleTcpClientSettings.cs
@@ -29,18 +29,53 @@ namespace SimpleTcp
         }
 
         /// <summary>
-        /// The number of seconds to wait when attempting to connect.
+        /// The number of milliseconds to wait when attempting to connect.
         /// </summary>
-        public int ConnectTimeoutSeconds
+        public int ConnectTimeoutMs
         {
             get
             {
-                return _ConnectTimeoutSeconds;
+                return _ConnectTimeoutMs;
             }
             set
             {
-                if (value < 1) throw new ArgumentException("ConnectTimeoutSeconds must be greater than zero.");
-                _ConnectTimeoutSeconds = value;
+                if (value < 1) throw new ArgumentException("ConnectTimeoutMs must be greater than zero.");
+                _ConnectTimeoutMs = value;
+            }
+        }
+
+        /// <summary>
+        /// Maximum amount of time to wait before considering the server to be idle and disconnecting from it. 
+        /// By default, this value is set to 0, which will never disconnect due to inactivity.
+        /// The timeout is reset any time a message is received from the server.
+        /// For instance, if you set this value to 30000, the client will disconnect if the server has not sent a message to the client within 30 seconds.
+        /// </summary>
+        public int IdleServerTimeoutMs
+        {
+            get
+            {
+                return _IdleServerTimeoutMs;
+            }
+            set
+            {
+                if (value < 0) throw new ArgumentException("IdleClientTimeoutMs must be zero or greater.");
+                _IdleServerTimeoutMs = value;
+            }
+        }
+
+        /// <summary>
+        /// Number of milliseconds to wait between each iteration of evaluating the server connection to see if the configured timeout interval has been exceeded.
+        /// </summary>
+        public int IdleServerEvaluationIntervalMs
+        {
+            get
+            {
+                return _IdleServerEvaluationIntervalMs;
+            }
+            set
+            {
+                if (value < 1) throw new ArgumentOutOfRangeException("IdleServerEvaluationIntervalMs must be one or greater.");
+                _IdleServerEvaluationIntervalMs = value;
             }
         }
 
@@ -59,7 +94,9 @@ namespace SimpleTcp
         #region Private-Members
 
         private int _StreamBufferSize = 65536;
-        private int _ConnectTimeoutSeconds = 5;
+        private int _ConnectTimeoutMs = 5000;
+        private int _IdleServerTimeoutMs = 0;
+        private int _IdleServerEvaluationIntervalMs = 1000;
 
         #endregion